# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ProfileResponse:
    properties:
      addresses:
        items:
          type: string
        type: array
      country:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      address:
        description: Phone      string                 `json:"phone_number" validate:"omitempty"`
        items:
          type: string
        type: array
      country:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        type: string
    required:
    - email
    - new_password
    type: object
...
  dto.UserUpdateRequest:
    properties:
      addresses:
        items:
          type: string
        type: array
      country:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for customer and merchant authentication in a multivendor platform
  license:
    name: MIT
  termsOfService: http://example.com/terms/
  title: Multivendor API
  version: "1.0"
paths:
...
  /customer/auth/google/callback:
    get:
      description: Processes Google OAuth callback and returns JWT token
      parameters:
      - description: OAuth code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: Code not provided
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Handle Google OAuth callback for customer
      tags:
      - Customer
  /customer/login:
    post:
      consumes:
      - application/json
      description: Authenticates a customer using email and password
      parameters:
      - description: Customer login credentials
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: Invalid request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Invalid role
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Customer login
      tags:
      - Customer
  /customer/logout:
    post:
      description: Invalidates the customer's JWT token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Authorization header required
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Customer logout
      tags:
      - Customer
  /customer/profile:
    get:
      description: Retrieves the customer's profile information
      produces:
      - application/json
      responses:
        "200":
          description: Profile details
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get customer profile
      tags:
      - Customer
  /customer/register:
    post:
      consumes:
      - application/json
      description: Creates a new customer account with email, name, password, and
        optional country
      parameters:
      - description: Customer registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: Invalid request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Register a new customer
      tags:
      - Customer
  /customer/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the customer's password (unprotected; add verification in
        production)
      parameters:
      - description: Reset details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Reset customer password
      tags:
      - Customer
  /customer/update:
    patch:
      description: Updates Customer profile
      parameters:
      - description: Update Customer Profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Authorization header required
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Customer Profile Update
      tags:
      - Customer
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"